Title: 입력받아 계산, Time: 113ms, Memory: 24MB, Status: Passed - CodeTree

+++ 240122/입력받아 계산/input-calculate.py
@@ -1,5 +1,5 @@
 a = input() #문자열로 입력받음
-a = float(a)
+a = int(a)
 print(a + 2)
 
 #a = int(input())
